[Code] added a test
