+++ x-pack/legacy/plugins/code/server/routes/lsp.ts
@@ -5,10 +5,8 @@
  */
 
 import Boom from 'boom';
-import { groupBy, last } from 'lodash';
 import { ResponseError } from 'vscode-jsonrpc';
 import { ResponseMessage } from 'vscode-jsonrpc/lib/messages';
-import { Location, Position } from 'vscode-languageserver-types';
 
 import { SymbolLocator } from '@elastic/lsp-extension';
 import {
@@ -18,22 +16,16 @@ import {
 } from '../../common/lsp_error_codes';
 import { parseLspUrl } from '../../common/uri_util';
 import { Logger } from '../log';
-import { CTAGS, GO } from '../lsp/language_servers';
 import { SymbolSearchClient } from '../search';
 import { CodeServerRouter } from '../security';
 import { ServerOptions } from '../server_options';
-import {
-  expandRanges,
-  extractSourceContent,
-  LineMapping,
-  mergeRanges,
-} from '../utils/composite_source_merger';
-import { detectLanguage } from '../utils/detect_language';
+
 import { EsClientWithRequest } from '../utils/esclient_with_request';
 import { promiseTimeout } from '../utils/timeout';
 import { RequestFacade, ResponseToolkitFacade } from '../..';
 import { CodeServices } from '../distributed/code_services';
 import { GitServiceDefinition, LspServiceDefinition } from '../distributed/apis';
+import { findTitleFromHover, groupFiles } from '../utils/lsp_utils';
 
 const LANG_SERVER_ERROR = 'language server error';
 
@@ -45,6 +37,7 @@ export function lspRoute(
   const log = new Logger(server.server);
   const lspService = codeServices.serviceFor(LspServiceDefinition);
   const gitService = codeServices.serviceFor(GitServiceDefinition);
+
   server.route({
     path: '/api/code/lsp/textDocument/{method}',
     async handler(req: RequestFacade, h: ResponseToolkitFacade) {
@@ -134,7 +127,10 @@ export function lspRoute(
           }
         }
       }
-      const files = await groupFiles(req, locations);
+      const files = await groupFiles(locations, async loc => {
+        const ep = await codeServices.locate(req, loc.uri);
+        return await gitService.blob(ep, loc);
+      });
       return { title, files, uri, position };
     },
   });
@@ -164,7 +160,10 @@ export function lspRoute(
           },
         });
         const title: string = await findTitleFromHover(hover, uri, position);
-        const files = await groupFiles(req, response.result);
+        const files = await groupFiles(response.result, async loc => {
+          const ep = await codeServices.locate(req, loc.uri);
+          return await gitService.blob(ep, loc);
+        });
         return { title, files, uri, position };
       } catch (error) {
         log.error(error);
@@ -184,88 +183,6 @@ export function lspRoute(
       }
     },
   });
-
-  async function groupFiles(req: RequestFacade, list: Location[]) {
-    const files = [];
-    const groupedLocations = groupBy(list, 'uri');
-    for (const url of Object.keys(groupedLocations)) {
-      const { repoUri, revision, file } = parseLspUrl(url)!;
-      const ep = await codeServices.locate(req, repoUri);
-      const locations: Location[] = groupedLocations[url];
-      const lines = locations.map(l => ({
-        startLine: l.range.start.line,
-        endLine: l.range.end.line,
-      }));
-      const ranges = expandRanges(lines, 1);
-      const mergedRanges = mergeRanges(ranges);
-      try {
-        const blob = await gitService.blob(ep, { uri: repoUri, path: file!, revision });
-        if (blob.content) {
-          const source = blob.content.split('\n');
-          const language = blob.lang;
-          const lineMappings = new LineMapping();
-          const code = extractSourceContent(mergedRanges, source, lineMappings).join('\n');
-          const lineNumbers = lineMappings.toStringArray();
-          const highlights = locations.map(l => {
-            const { start, end } = l.range;
-            const startLineNumber = lineMappings.lineNumber(start.line);
-            const endLineNumber = lineMappings.lineNumber(end.line);
-            return {
-              startLineNumber,
-              startColumn: start.character + 1,
-              endLineNumber,
-              endColumn: end.character + 1,
-            };
-          });
-          files.push({
-            repo: repoUri,
-            file,
-            language,
-            uri: url,
-            revision,
-            code,
-            lineNumbers,
-            highlights,
-          });
-        }
-      } catch (e) {
-        // can't load this file, ignore this result
-      }
-    }
-    return groupBy(files, 'repo');
-  }
-
-  async function findTitleFromHover(hover: ResponseMessage, uri: string, position: Position) {
-    let title: string;
-    if (hover.result && hover.result.contents) {
-      if (Array.isArray(hover.result.contents)) {
-        const content = hover.result.contents[0];
-        title = hover.result.contents[0].value;
-        const lang = await detectLanguage(uri.replace('file://', ''));
-        // TODO(henrywong) Find a gernal approach to construct the reference title.
-        if (content.kind) {
-          // The format of the hover result is 'MarkupContent', extract appropriate pieces as the references title.
-          if (GO.languages.includes(lang)) {
-            title = title.substring(title.indexOf('```go\n') + 5, title.lastIndexOf('\n```'));
-            if (title.includes('{\n')) {
-              title = title.substring(0, title.indexOf('{\n'));
-            }
-          }
-        } else if (CTAGS.languages.includes(lang)) {
-          // There are language servers may provide hover results with markdown syntax, like ctags-langserver,
-          // extract the plain text.
-          if (title.substring(0, 2) === '**' && title.includes('**\n')) {
-            title = title.substring(title.indexOf('**\n') + 3);
-          }
-        }
-      } else {
-        title = hover.result.contents as 'string';
-      }
-    } else {
-      title = last(uri.split('/')) + `(${position.line}, ${position.character})`;
-    }
-    return title;
-  }
 }
 
 export function symbolByQnameRoute(router: CodeServerRouter, log: Logger) {
